Provider tests: add tests for integer types

+++ Json/Fliox.Hub.SQLServer/SQLServerExtensions.cs
@@ -34,7 +34,7 @@ namespace Friflo.Json.Fliox.Hub.SQLServer
         
         internal static string GetSqlType(StandardTypeId typeId) {
             switch (typeId) {
-                case StandardTypeId.Uint8:      return "tinyint";
+                case StandardTypeId.Uint8:      return "smallint"; // no byte type (0-255) available
                 case StandardTypeId.Int16:      return "smallint";
                 case StandardTypeId.Int32:      return "int";
                 case StandardTypeId.Int64:      return "bigint";
